cmake: Overally cleanup/testing update
This brings SPIRV-HEADERS up to date with Vulkan-Headers in terms
of modern CMake practices/testing.

Minimimum is raised to 3.14 for ARCH_INDEPENDENT functionality,
which helps reduces installation of header only libraries. And
makes writing/maintaining the CMake much easier.

Lots of testing was added to ensure the stability/interface for
package managers. In particular the pkgconfig support is now
tested much more extensively for correctness.

The CMake minimum and add_subdirectory workflow is now also
explicitly tested for compatibility.

closes #282
closes #324
closes #325