Simplify enabling capabilities for image Dim
A lot of these capabilities implied another of the enabling capabilities
so didn't really serve their intended purpose. Remove them to simplify
this section of the grammar.

This affects one table in the spec and a corresponding spec change will
clarify that this makes no change to the meaning or scope of these
capabilities.