Add "aliases" fields to the grammar
Promotion of extensions lead to duplicated entries.
The patch introduce a new entry "aliases" to the grammar.
It is a list of strings, where each string is an alternative name for the token,
and applies on instructions and operand kind's enumerants.
The grammar is updated to use the fields and all duplicated were removed.

The header generator now processes the aliases when generating the headers to account for the change.
This ensure the generated headers doesn't change.