Add definitions for TOSA.001000.1 extended instruction set (#526)
Also support custom header file names and header guard names in
generated headers. This is useful when multiple extended instruction
sets share a prefix. In the case of TOSA SPIR-V, multiple versions
exist but we want all definitions to be prefixed with "TOSA" but
have their own header and a different header guard such that it is
possible to use the definitions in code that deals with multiple
versions of the instruction set.

Change-Id: I7c18b72fd229e586da7cb1070d24cfcd69826827

Signed-off-by: Kevin Petit <kevin.petit@arm.com>